+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 Changelog for package labelizer
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+0.1.2 (2017-11-17)
+------------------
+* GUI can now display sample images
+* Contributors: Florian Jordan
+
 0.1.1 (2017-11-17)
 ------------------
 * catches wrong storage of .png as .jpg